requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
neo4j
sentence-transformers
google-generativeai
networkx
scikit-learn
numpy
apscheduler
